adding of category names in the last five tricks
