+++ src/Controller/HomeController.php
@@ -13,7 +13,7 @@ namespace App\Controller;
 
 use App\Model\CategoryManager;
 use App\Model\TricksManager;
-use App\Model\TricksCategManager;
+
 
 class HomeController extends AbstractController
 {
@@ -29,8 +29,8 @@ class HomeController extends AbstractController
     {
         $categoryManager = new CategoryManager();
         $categories = $categoryManager->selectAll('name');
-        $tricksCategManager = new TricksCategManager();
-        $lastFiveTricks = $tricksCategManager->selectJoin('tricks_id', 'DESC', '5');
+        $tricksManager = new TricksManager();
+        $lastFiveTricks = $tricksManager->selectJoin('tricks_id', 'DESC', '5');
         return $this->twig->render(
             'Home/index.html.twig',
             ['categories' => $categories, 'lastFiveTricks' => $lastFiveTricks]
